created floattable namespace fixed some warnings
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/mmstd_datatools/trunk@114 f95d2973-9550-45a5-b795-f9f66c6b6f94

+++ src/MultiParticleRelister.cpp
@@ -135,8 +135,8 @@ void MultiParticleRelister::copyData(const core::moldyn::SimpleSphericalParticle
     data.resize(relist.TargetListCount());
 
     for (size_t pi = 0; pi < inData.GetCount(); ++pi, colDat += colStep, verDat += verStep, relistData++) {
-        for (int i = 0; i < verSize; ++i) data[*relistData].push_back(verDat[i]);
-        for (int i = 0; i < colSize; ++i) data[*relistData].push_back(colDat[i]);
+        for (size_t i = 0; i < verSize; ++i) data[*relistData].push_back(verDat[i]);
+        for (size_t i = 0; i < colSize; ++i) data[*relistData].push_back(colDat[i]);
     }
 
 }
